Allow for arbitrary comparison operators
At present conditionize can only conditionally show elements when a
value is exactly equal to a specific control value.

This change allows for the conditions to be based on all other
comparison operators. Specifying the 'data-cond-operator' attribute as
one of '==', '!=', '<=', '<', '>=', '>' tells conditionize the operator
which should be used to make the comparison. When cond-operator is not
specified, we fall back to '=='.